+++ lib/src/main/java/edu/utexas/tacc/tapis/files/lib/caches/SystemsCacheNoAuth.java
@@ -88,6 +88,11 @@ public class SystemsCacheNoAuth
     }
   }
 
+  public void invalidateEntry(String tenantId, String systemId, String tapisUser) {
+    SystemCacheKey key = new SystemCacheKey(tenantId, systemId, tapisUser);
+    cache.invalidate(key);
+  }
+
   // ====================================================================================
   // =======  Private Classes ===========================================================
   // ====================================================================================